requirements.txt
alembic==1.8.1
anyio==3.6.1
asgiref==3.5.2
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.0
click==7.1.2
click-didyoumean==0.3.0
cls-client==1.4.0
dnspython==2.2.1
email-validator==1.2.1
fastapi==0.79.0
FastAPI-SQLAlchemy==0.2.1
fastapi-versioning==0.10.0
google-api-core==2.8.2
google-auth==2.10.0
google-cloud-secret-manager==2.12.2
googleapis-common-protos==1.56.4
greenlet==1.1.3
grpc-google-iam-v1==0.12.4
grpcio==1.47.0
grpcio-status==1.47.0
h11==0.13.0
httptools==0.4.0
idna==3.3
importlib-metadata==4.12.0
importlib-resources==5.9.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.3
MarkupSafe==2.1.1
orjson==3.7.11
p==1.3.0
passlib==1.7.4
proto-plus==1.22.0
protobuf==3.20.1
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.9.2
PyJWT==2.4.0
python-dotenv==0.20.0
python-multipart==0.0.5
PyYAML==6.0
requests==2.28.1
rsa==4.9
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.41
starlette==0.19.1
typing-extensions==4.3.0
ujson==5.4.0
urllib3==1.26.11
uvicorn==0.17.6
uvloop==0.16.0
watchgod==0.8.2
websockets==10.3
zipp==3.8.1